move command and validators from application to domain layer
