+++ src/main/java/com/pafolder/librarian/domain/command/Command.java
@@ -1,4 +1,4 @@
-package com.pafolder.librarian.application.command;
+package com.pafolder.librarian.domain.command;
 
 @FunctionalInterface
 public interface Command<T> {
