+++ src/main/java/com/pafolder/librarian/domain/validator/BookingLimit.java
@@ -1,4 +1,4 @@
-package com.pafolder.librarian.application.validator;
+package com.pafolder.librarian.domain.validator;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
@@ -9,13 +9,16 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Documented
-@Constraint(validatedBy = ViolationLimitValidator.class)
-@Target( { ElementType.METHOD, ElementType.FIELD })
+@Constraint(validatedBy = BookingLimitValidator.class)
+@Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface ViolationLimit {
-  String message() default "Borrowing is prohibited because the violation limit exceeded";
-  int maxViolation() default 1;
-  int borrowDurationDays() default 14;
+public @interface BookingLimit {
+
+  int maxBooksAllowedAtOnce() default 3;
+
+  String message() default "Borrowing is prohibited because the limit of {maxBooksAllowedAtOnce} books reached";
+
   Class<?>[] groups() default {};
+
   Class<? extends Payload>[] payload() default {};
 }
