+++ src/main/java/com/pafolder/librarian/infrastructure/controller/view/UserView.java
@@ -4,16 +4,13 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.pafolder.librarian.domain.model.User;
 import com.pafolder.librarian.domain.model.User.Role;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 
 @JsonInclude(Include.NON_NULL)
-@AllArgsConstructor
-@NoArgsConstructor
+@RequiredArgsConstructor
 public class UserView {
 
-  private User user;
+  private final User user;
 
   public Integer getId() {
     return user.getId();
